projects/Facebook-Script-Extract/requirements.txt
yt-dlp>=2024.4.9
openai>=1.30.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
celery>=5.3.0
redis>=5.0.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
websockets>=12.0
playwright>=1.43.0
requests>=2.31.0
